+++ benchmark/fashionmnist/fashionmnist_batched_test.go
@@ -58,7 +58,7 @@ func TestFashionMNISTBatched(t *testing.T) {
 				continue
 			}
 			wf := (rng.Float64()*2.0 - 1.0) * initWeightMax
-			w := int32(math.Round(wf * intScale))
+			w := int64(math.Round(wf * intScale))
 			if w == 0 {
 				w = 1
 			}
@@ -73,7 +73,7 @@ func TestFashionMNISTBatched(t *testing.T) {
 				continue
 			}
 			wf := (rng.Float64()*2.0 - 1.0) * initWeightMax
-			w := int32(math.Round(wf * intScale))
+			w := int64(math.Round(wf * intScale))
 			if w == 0 {
 				w = 1
 			}
