+++ benchmark/mnist/mnist_activity_512_test.go
@@ -52,7 +52,7 @@ func TestMNISTActivity512(t *testing.T) {
 				continue
 			}
 			wf := (rng.Float64()*2.0 - 1.0) * initWeightMax
-			w := int32(math.Round(wf * intScale))
+			w := int64(math.Round(wf * intScale))
 			if w == 0 {
 				w = 1
 			}
@@ -67,7 +67,7 @@ func TestMNISTActivity512(t *testing.T) {
 				continue
 			}
 			wf := (rng.Float64()*2.0 - 1.0) * initWeightMax
-			w := int32(math.Round(wf * intScale))
+			w := int64(math.Round(wf * intScale))
 			if w == 0 {
 				w = 1
 			}
