+++ benchmark/mnist/mnist_tuned_test.go
@@ -55,7 +55,7 @@ func TestMNISTTuned(t *testing.T) {
 				continue
 			}
 			wf := (rng.Float64()*2.0 - 1.0) * initWeightMax
-			w := int32(math.Round(wf * intScale))
+			w := int64(math.Round(wf * intScale))
 			if w == 0 {
 				w = 1
 			}
@@ -70,7 +70,7 @@ func TestMNISTTuned(t *testing.T) {
 				continue
 			}
 			wf := (rng.Float64()*2.0 - 1.0) * initWeightMax
-			w := int32(math.Round(wf * intScale))
+			w := int64(math.Round(wf * intScale))
 			if w == 0 {
 				w = 1
 			}
